refactor: 엔티티 클래스의 패키지 경로 재배치
