--- src/main/kotlin/goodspace/bllsoneshot/entity/Assignment.kt
@@ -1,14 +1,0 @@
-package goodspace.bllsoneshot.entity
-
-import jakarta.persistence.*
-
-@Entity
-class Assignment(
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    val mentee: User
-) {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null
-}
