--- src/main/kotlin/goodspace/bllsoneshot/entity/Counseling.kt
@@ -1,17 +1,0 @@
-package goodspace.bllsoneshot.entity
-
-import jakarta.persistence.*
-
-@Entity
-class Counseling(
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    val mentor: User,
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    val mentee: User
-) {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null
-}
