--- src/main/kotlin/goodspace/bllsoneshot/entity/ProofShot.kt
@@ -1,14 +1,0 @@
-package goodspace.bllsoneshot.entity
-
-import jakarta.persistence.*
-
-@Entity
-class ProofShot(
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(nullable = false)
-    val assignment: Assignment
-) {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null
-}
