--- src/main/kotlin/goodspace/bllsoneshot/entity/User.kt
@@ -1,18 +1,0 @@
-package goodspace.bllsoneshot.entity
-
-import jakarta.persistence.*
-
-@Entity
-@Table(name = "users")
-class User(
-    @ManyToOne(fetch = FetchType.LAZY)
-    val mentor: User? = null,
-
-    val role: UserRole
-) {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null
-
-    var refreshToken: String? = null
-}
