+++ src/main/kotlin/goodspace/bllsoneshot/entity/assignment/Planner.kt
@@ -1,0 +1,21 @@
+package goodspace.bllsoneshot.entity.assignment
+
+import goodspace.bllsoneshot.entity.user.User
+import jakarta.persistence.Entity
+import jakarta.persistence.FetchType
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.ManyToOne
+
+@Entity
+class Planner(
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(nullable = false)
+    val mentee: User
+) {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long? = null
+}
