+++ src/main/kotlin/goodspace/bllsoneshot/entity/user/UserRole.kt
@@ -1,4 +1,4 @@
-package goodspace.bllsoneshot.entity
+package goodspace.bllsoneshot.entity.user
 
 enum class UserRole {
     ROLE_MENTOR, ROLE_MENTEE
